Better handling of implicit targets
This is the name I'm giving to *files* that are depended on as if
they are targets, but which are never declared as targets in the
file.  We create entries for these when reading the maker file, but
set them up so that they cannot be built (as there is no rule).
A warning will still be raised if these are declared but the files do
not exist -- later I might make that warning suppressable (for the case
where files will be generated by some other rule).