Refactor 'targets' to be self contained.
There's a slightly artificial distinction now between "active" and
"inactive" targets, but it's just for bookkeeping.  The maker instance
passes itself through to each target so that they can run themselves.
There's a bit more tidying to go, but this isolates all of the major
switching between behaviours into the targets.R file and class.