Re-read source files when rebuilding if files have changed
If any of the source files that we included (either directly or via a
directory) are changed, then re-build the environment before re-running
make.  This does not pick up files sourced by make though; getting
that is a bit harder, but probably doable.

Still need to do the same to trigger re-sourcing the makerfile.