Support for chained rules.
These will be useful if there are a "pipeline" of targets, but the
intermediate steps aren't really interesting enough to warrant their
own target names.  See tests/testthatain.yml for the syntax: instead
of having sections 'rule' and 'depends', a target can contain a
section 'chained'.  That section is an *ordered* list, each
element of which is a yaml dictionary with elements rule, depends,
cleanup_level.  This builds a series of implicit targets.