Generate plain R scripts that can run a project without maker
This is just a first stab at this.  Chained targets are not supported
(because they use non-syntactic names) and there is no commenting.
It would be nice if the structure implied by fake targets could be
used here to add "sections".  However, that's not how the topological
sort actually plays out.  In any case, for simple cases this will
hopefully demystify what is going on.