Basic knitr support.
See tests/testthat/knitr.yml for the syntax.  The input file is
inferred from the output file (Rmd -> md) and Rnw is not currently
supported, though that won't be hard because knitr does all the hard work
for us.  The `quiet` flag here will be useful: possibly enough to
make it true by default.  Objects that are depended upon will be
exported to the environment that knitr uses.  It might be useful to
broaden that out to export all objects, but that gets tricky because they
might not all be made.  Better would probably be to allow depending
on a fake target and getting all that targets dependencies.  All
functions that are included via the `sources:` field are available as
knitr will run in an environment that has the store environment as a parent.

cc: @dfalster