Support for loading per-target packages
Fixes #12, mostly

I've avoided the case where a package can be loaded for a fake target
(being available for all its dependents) because that might not correspond
to the build order: that depends on the order of the topological sort.
Parallel builds will make that even harder!

Eventually what I think would cover the most common use case, is if declaring
an extra package in a fake (i.e. grouping) target would add it to all its
dependents.  That matches for the specific case causing problems for #12.