Support for literal arguments
If a given argument is logical (TRUE/FALSE), integer, numeric or
complex, then it will be interpreted as literal arguments automatically.
For string arguments or names of variables that are defined within R's
global/base environments (e.g., pi) or within the source files, the
argument must be escaped with `I()` - e.g., I(variable_name) or
I("literal string constant").  The *values* of variables are stored
in makers database, so changing these will trigger a rebuild.

Fixes #17