Support functions for working with active bindings
The idea is that active bindings dropped off in the main environment
act as a read-only interface to maker; accessing a variable causes
just-in-time resolution of the variable.  Assigning to a variable
is an error, and it will always be checked to see if it is up to date.

That implies a significant overhead (on top of the nontrivial overhead
that active binding functions incur over plain lookup) so there's a
"resolve" function that replaces all bindings with finalised objects.
That's probably not the best idea in general because it *will* make
everything.