Check that files are correct cases.
Mac and Windows users can mix up cases (especially file.r vs
file.R) without real penalty, but it causes issues for
portability.  Especially if people are going to run things
on travis or other Linux based systems, getting this right at
the outset is a good idea.  This implements a couple of hacks
to try and determine if a file would not exist on a system
that is case sensitive, and to work out what the canonical
form of the filename is.