Port object storage to storr
Possible speedup here, though expect archives to get a bit bloated
until the automatic garbage collection is worked out.  The main
advantage is a great simplification of the code here, and separating
out an idea that might be useful elsewhere.  The other thing this
paves the way for is support of list targets (list support is
native to storr) and database support (Redis support is present in
storr) plus easy translation between the different backends.  An
"envirment" backend could also be used for testing as it will be
heaps faster.