Better error messages when file downloading fails.
This appends the underlying error to the general "downloading failed"
error, which gives an informative error message in the case where
downloader is not installed and helps diagnose the root cause.  This might
be eventually replaced by using R's internal download support now that
it supports https URLs to some degree.

Fixes #57