New option allow_missing_packages, available on most api fns
This allows users to force proceeding even if some packages are
missing.  This might result in errors in user code, so it's not
really ideal, but I have a use case for it already, so it might
be generally useful.  Missing packages are skipped during package
loading, no warning or information about this is given.